+++ src/map/ai/states/range_state.cpp
@@ -55,7 +55,7 @@ CRangeState::CRangeState(CCharEntity* PEntity, uint16 targid) :
         if (dsprand::GetRandomNumber(100) < chance)
         {
             //reduce delay by 10%-50%
-            delay = delay * (10 - dsprand::GetRandomNumber(1, 6)) / 10.f;
+            delay = (int16)(delay * (10 - dsprand::GetRandomNumber(1, 6)) / 10.f);
             m_rapidShot = true;
         }
     }
